chore(io-core): drop eight zero-consumer modules
reader, writer, bufreader_optimizer, shared_memory, direct_io, timeout_wrapper, io_priority_queue, and scheduler had no caller outside io-core's own tests and example. The scheduling algorithm and request timeout wrapper RustFS actually runs live in rustfs/src/storage; io-core keeps the config shapes they project into.

OperationProgress, the one item with a real consumer via rustfs-concurrency, moves to a new progress module and keeps its export path. memmap2 was only used by reader.rs and is dropped. Crate docs and CHANGELOG follow the removed APIs.
